requirements_tmzd.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
appdirs==1.4.4
asttokens==3.0.0; python_version >= '3.8'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
comm==0.2.2; python_version >= '3.8'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
debugpy==1.8.14; python_version >= '3.8'
decorator==5.2.1; python_version >= '3.8'
deprecation==2.1.0
distro==1.9.0; python_version >= '3.6'
dnspython==2.7.0; python_version >= '3.9'
dotenv==0.9.9
dynaconf==3.2.10; python_version >= '3.8'
email-validator==2.2.0; python_version >= '3.8'
evidently==0.6.7; python_version >= '3.8'
executing==2.2.0; python_version >= '3.8'
faker==37.1.0; python_version >= '3.9'
fastapi==0.115.12; python_version >= '3.8'
fastapi-cli==0.0.7; python_version >= '3.8'
fastembed==0.6.1; python_full_version >= '3.9.0'
filelock==3.18.0; python_version >= '3.9'
flatbuffers==25.2.10
fsspec==2025.3.2; python_version >= '3.9'
google-auth==2.39.0; python_version >= '3.7'
google-genai==1.11.0; python_version >= '3.9'
grpcio==1.71.0; python_version >= '3.9'
h11==0.14.0; python_version >= '3.7'
h2==4.2.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.8; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx[http2]==0.28.1; python_version >= '3.8'
huggingface-hub==0.30.2; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
ipykernel==6.29.5; python_version >= '3.8'
ipython==9.1.0; python_version >= '3.11'
ipython-pygments-lexers==1.1.1; python_version >= '3.8'
iterative-telemetry==0.0.10; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jupyter-client==8.6.3; python_version >= '3.8'
jupyter-core==5.7.2; python_version >= '3.8'
langchain-core==0.3.55; python_version >= '3.9' and python_version < '4.0'
langgraph==0.3.34; python_full_version >= '3.9.0' and python_version < '4.0'
langgraph-checkpoint==2.0.24; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-prebuilt==0.1.8; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-sdk==0.1.63; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langsmith==0.3.33; python_version >= '3.9' and python_version < '4.0'
litestar==2.15.2; python_version >= '3.8' and python_version < '4.0'
litestar-htmx==0.4.1; python_version >= '3.8' and python_version < '4.0'
loguru==0.7.3; python_version >= '3.5' and python_version < '4.0'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mmh3==5.1.0; python_version >= '3.9'
mpmath==1.3.0
msgspec==0.19.0; python_version >= '3.9'
multidict==6.4.3; python_version >= '3.9'
multipart==1.2.1; python_version >= '3.8'
mypy-extensions==1.1.0; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.4.2; python_version >= '3.10'
nltk==3.9.1; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
onnxruntime==1.21.1; python_version >= '3.10'
orjson==3.10.16; python_version >= '3.9'
ormsgpack==1.9.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pandas[parquet]==2.2.3; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
patsy==1.0.1; python_version >= '3.6'
pillow==11.2.1; python_version >= '3.9'
pip==25.0.1; python_version >= '3.8'
pipdeptree==2.26.1; python_version >= '3.9'
platformdirs==4.3.7; python_version >= '3.9'
plotly==5.24.1; python_version >= '3.8'
polyfactory==2.21.0; python_version >= '3.8' and python_version < '4.0'
portalocker==2.10.1; python_version >= '3.8'
prompt-toolkit==3.0.51; python_version >= '3.8'
protobuf==6.30.2; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
psycopg2-binary==2.9.10; python_version >= '3.8'
pure-eval==0.2.3
py-rust-stemmers==0.1.5
pyarrow==19.0.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.3; python_version >= '3.9'
pydantic-core==2.33.1; python_version >= '3.9'
pygments==2.19.1; python_version >= '3.8'
pyreadline3==3.5.4; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.2
pywin32==310
pyyaml==6.0.2; python_version >= '3.8'
pyzmq==26.4.0; python_version >= '3.8'
qdrant-client==1.14.2; python_version >= '3.9'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==14.0.0; python_full_version >= '3.8.0'
rich-click==1.8.8; python_version >= '3.7'
rich-toolkit==0.14.3; python_version >= '3.8'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
safetensors==0.5.3; python_version >= '3.7'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.2; python_version >= '3.10'
sentence-transformers==4.1.0; python_version >= '3.9'
setuptools==79.0.1; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
stack-data==0.6.3
starlette==0.46.2; python_version >= '3.9'
statsmodels==0.14.4; python_version >= '3.9'
sympy==1.13.3; python_version >= '3.8'
tenacity==9.1.2; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenizers==0.21.1; python_version >= '3.9'
torch==2.7.0; python_full_version >= '3.9.0'
tornado==6.4.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
transformers==4.51.3; python_full_version >= '3.9.0'
typer==0.15.2; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspect==0.9.0
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
ujson==5.10.0; python_version >= '3.8'
urllib3==2.4.0; python_version >= '3.9'
uuid6==2024.7.10; python_version >= '3.8'
uvicorn[standard]==0.34.2; python_version >= '3.9'
watchdog==6.0.0; python_version >= '3.9'
watchfiles==1.0.5; python_version >= '3.9'
wcwidth==0.2.13
websockets==15.0.1; python_version >= '3.9'
win32-setctime==1.2.0; python_version >= '3.5'
xxhash==3.5.0; python_version >= '3.7'
zstandard==0.23.0; python_version >= '3.8'
